+++ src/uikit/components/Badge/Badge.tsx
@@ -2,7 +2,7 @@ import type { ColorKey } from '../../theme/types';
 import React from 'react';
 import { View } from 'react-native';
 
-import useTheme from '../../hooks/useTheme';
+import useTheme from '../../hooks/useAppTheme';
 import Text from '../Text/Text';
 
 import styles from './Badge.style';
