+++ src/uikit/components/Button/Button.tsx
@@ -2,7 +2,7 @@ import type { Sizes } from '../../theme/types';
 import React from 'react';
 import { Text, Pressable, View } from 'react-native';
 
-import useTheme from '../../hooks/useTheme';
+import useTheme from '../../hooks/useAppTheme';
 import Icon from '../Icon/Icon';
 import type { IconComponentName } from '../Icon/IconComponents';
 import Spinner from '../Spinner/Spinner';
