+++ src/uikit/components/Icon/Icon.tsx
@@ -2,7 +2,7 @@ import type { ColorKey, Sizes } from '../../theme/types';
 import Color from 'color';
 import React, { useMemo } from 'react';
 
-import useTheme from '../../hooks/useTheme';
+import useTheme from '../../hooks/useAppTheme';
 
 import Icons from './IconComponents';
 import type { IconComponentName } from './IconComponents';
