+++ src/uikit/components/Indicators/QualityIndicator/QualityIndicator.tsx
@@ -2,7 +2,7 @@ import type { ColorKey } from '../../../theme/types';
 import React from 'react';
 import { View } from 'react-native';
 
-import useTheme from '../../../hooks/useTheme';
+import useTheme from '../../../hooks/useAppTheme';
 
 import styles from './QualityIndicator.style';
