+++ src/uikit/components/Pill/Pill.tsx
@@ -3,7 +3,7 @@ import type { Sizes } from '../../theme/types';
 import React from 'react';
 import { View } from 'react-native';
 
-import useTheme from '../../hooks/useTheme';
+import useTheme from '../../hooks/useAppTheme';
 import Text from '../Text/Text';
 
 import styles from './Pill.style';
